+++ bmnclient/resources/translations/de_DE.py.ts
@@ -3,66 +3,120 @@
 <context>
     <name>@default</name>
     <message>
-        <location filename="fiat_rate.py" line="130"/>
+        <location filename="fiat_rate.py" line="129"/>
         <source>-- None --</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="currency.py" line="27"/>
+        <location filename="currency.py" line="26"/>
         <source>US Dollar</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="currency.py" line="32"/>
+        <location filename="currency.py" line="31"/>
         <source>Euro</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="password.py" line="18"/>
+        <source>Horrible</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="password.py" line="19"/>
+        <source>Weak</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="password.py" line="20"/>
+        <source>Medium</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="password.py" line="21"/>
+        <source>Good</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="password.py" line="22"/>
+        <source>Strong</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="password.py" line="23"/>
+        <source>Paranoiac</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="key_store.py" line="33"/>
+        <source>Unknown Key Store error.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="key_store.py" line="73"/>
+        <source>This will destroy all saved information and you can lose your money!
+Please make sure you remember the Seed Phrase.
+
+Reset?</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="key_store.py" line="97"/>
+        <source>Wrong Key Store password.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="key_store.py" line="137"/>
+        <source>Wrong Seed Phrase.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="address.py" line="171"/>
+        <source>New {coin_name} transaction</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="address.py" line="175"/>
+        <source>{tx_name}
+{amount} {unit} / {fiat_amount} {fiat_unit}</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
 <context>
     <name>KeyStore</name>
+</context>
+<context>
+    <name>PasswordDialog</name>
     <message>
-        <location filename="key_store.py" line="213"/>
-        <source>Wrong password.</source>
+        <location filename="key_store.py" line="319"/>
+        <source>Wrong Key Store password.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="key_store.py" line="216"/>
-        <source>Seed phrase not found.</source>
+        <location filename="key_store.py" line="322"/>
+        <source>Seed Phrase not found.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="key_store.py" line="325"/>
+        <source>Unknown Key Store error.</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
 <context>
     <name>SystemTrayIcon</name>
     <message>
-        <location filename="system_tray.py" line="54"/>
-        <source>Show</source>
+        <location filename="system_tray.py" line="52"/>
+        <source>&amp;Show/Hide main window</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="system_tray.py" line="58"/>
-        <source>Hide</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="system_tray.py" line="65"/>
-        <source>Quit</source>
+        <location filename="system_tray.py" line="57"/>
+        <source>&amp;Quit</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
 <context>
     <name>UIManager</name>
-    <message>
-        <location filename="ui_manager.py" line="147"/>
-        <source>Text &apos;%s&apos; copied to clipboard</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="ui_manager.py" line="91"/>
-        <source>New transaction
-Coin: {coin_name}
-ID: {tx_name}
-Amount: {amount} {unit} / {fiat_amount} {fiat_unit}</source>
-        <translation type="unfinished"></translation>
-    </message>
 </context>
 </TS>